Revise arguments to be an ordered list
Positional parameters depend on order. Representing them as a hashmap
destroys this order. It enables nonsense such as optional parameters
followed by required parameters, with no way to fix it.

It's less attractive, but the obvious solution (and the one implemented
by clap) is to represent them as a list.